ImageEditor - remove dark corners on rotation
